Added Account support framework
Also fixed a bigger mistake I made earlier. It is also solely a framework as of now and doesn't do anything other than sitting in there, I'll create all the integrations later.
There's still a lot of work to do but I'm confident that it will be done fairly soon. The two databases way of thinking on how the Economy storage works is probably not ideal but It's the best thing that we can get without rewriting the entire plugin.

Signed-off-by: Geolykt <admin@geolykt.de>