Cleanup and version bumps
Now using Java 11, because people should not use older versions and we personally do not use Java 11. Also changed some other stuff to enforce the fact that this is a fork that will probably be an independent project.
Also broke support with Java 8, because why wouldn't I? On another notice I also included changes that I wanted to commit before, but didn't get commited for some reason.
I might also change the packet name, but for now It's going to stay to be `dev.wwst.easyconomy`

Signed-off-by: Geolykt <admin@geolykt.de>