Removed even more static abuse
This should be the last commit on that particular topic and honestly was a bit harder than I expected it to be. Next commit will rework permissions (which I already changed somewhat in this commit, but oh well) for command execution.
As a downside of this commit, Configuration upgrading no longer works than it was, but since this branch is meant for personal production use, I don't really care here.

Signed-off-by: Geolykt <admin@geolykt.de>