Allow for external compression of backups
Technically it allows for pretty much everything when a backup is performed, so yeah. It should be noted that this was a real pain to get an example working, but after many hours of trail-and-error I can finally present one. Of course the example is commented out since it only works on linux (maybe also on other OSes, but not on Windos). This was such a pain to make and probably not worth it but eh :shrug:.

Signed-off-by: Geolykt <admin@geolykt.de>