Thread safety for the BinaryDataStorage and added v2 complicance
The new v2 storage format offers marginal compression by not saving the balance of accounts that have a predictable balance (i. e. those that have only joined, but not actually played on the server and as such have a balance equal to the default). This is basically a compromise between unpurged files and files where everyone that has a default balance is ommited from, which while technically even more effecive (would save 24 bytes per player with v1 and 16 bytes per player compared to v2) is more prone to strange events if the default balance would be altered. V1 and V2 are somewhat interchangeable, and a v1 parser would either be able to fully functionally read the database or fail at the basic unit test. Of course after a v1 parser reads a v2 database the data is has read will become polluted, but this isn't all too much of an issue as the original important data is still preserved, but inactive accounts may get signficant more money than they should get (this is a 1 in 3 chance, so still a bit high), but since this usually results in the player either having impossibly small or really high amounts of money, this could be caught by an addon and cleaned.

Signed-off-by: Geolykt <admin@geolykt.de>