Big refractors: modularised the project
This was defenitely made with a Minestom port in mind, but of course a port to forge and fabric would also be possible, although a lot harder as the both use gradle even more than Minestom for example.
It was also made to make Bukkit-indepent stuff free from the chains of GPL, so anyone using the api would be able to freely choose what they use as a license. While I'm kindof unsure on how legal that just was, most
of the interfaces were made by me (well, actually all of them), so I'll give myself the permission to do so either way. The only issue that the new format is making is that the bank storage engine is responsible of decoding data while the bank implementation is responsible for encoding them, with minimal informational interchange between them, which makes them to be forced to agree on one common (which is currently implemented) encoding format.
This hopefully is also the last big refractor, while the only thing that could be refractored is the thing I noted above and maybe feedback on how data is processed.

Signed-off-by: Geolykt <admin@geolykt.de>