Added early Minestom support
This is basically a headless economy implementation, not even implementing any commands, which will likely be changed soon.
On another notice some concepts we are making use of (i. e. flatfile storage) are not really implementable with our API, so maybe an extension needs to be added to allow for these concepts.

The commit also fixes a compilation deadlock with the bukkit module by updating the maven-shading-plugin to an acceptable version.

Signed-off-by: Geolykt <admin@geolykt.de>