Added balance and givemoney commands
The givemoney command can be executed by anyone right now, which we'll change later down the line to allow for proper testing, perhaps the command will be temporarily removed after that since permissions are a joke.
There's also a bigfix included in this build when the format method of the DefaultEconomyProvider is called.

Known bugs: the plugin does not save correctly.

Signed-off-by: Geolykt <admin@geolykt.de>